emailsummarizer-main/emailsummarizer-main/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-auth-oauthlib
google-api-python-client
pydantic
SQLAlchemy==1.4.32
alembic==1.8.1
